refactor: with TimelineList
