+++ src/web/src/components/index.ts
@@ -10,3 +10,4 @@ export { default as NotSupportText } from './NotSupportText';
 export { default as StepTitle } from './StepTitle';
 export { default as TimelineItemBox } from './TimelineItemBox';
 export { default as TempSavedPostModal } from './TempSavedPostModal';
+export { default as TimelineItemList } from './TimelineItemList';
