+++ src/web/src/pages/HomePage.tsx
@@ -1,19 +1,10 @@
 import { useState } from 'react';
-import { useNavigate } from '@tanstack/react-router';
 
-import { Tabs, Header, ContentLayout, TimelineItemBox } from '@/components';
-import {
-  ReplyBottomSheet,
-  ShareBottomSheet,
-  ViewsBottomSheet,
-} from '@/components/bottomSheet';
-import { usePaintAction } from '@/hooks';
 import type { TimelineItem } from '@/@types';
 import { createDummyTimelineItem } from '@/utils';
+import { Tabs, Header, ContentLayout, TimelineItemList } from '@/components';
 
 function HomePage() {
-  const navigate = useNavigate();
-  const paintAction = usePaintAction();
   const [paints] = useState<TimelineItem[]>(() => createDummyTimelineItem(10));
 
   return (
@@ -38,80 +29,21 @@ function HomePage() {
           {
             label: '추천',
             content: (
-              <ContentLayout
-                className="flex flex-col gap-[12px] pl-[12px] pr-[4px] mt-0 pb-[50px] max-h-[calc(100%-94px)] divide-y divide-blueGrey-400"
-                onScroll={paintAction.onScrollLayout}
-              >
-                {paints.map((paint) => (
-                  <TimelineItemBox
-                    key={paint.id}
-                    item={paint}
-                    className="pt-[10px]"
-                    isShowMenu={
-                      paintAction.isShowMoreMenu.id === paint.id &&
-                      paintAction.isShowMoreMenu.show
-                    }
-                    onClickReply={() =>
-                      navigate({
-                        to: '/post/edit',
-                        search: { postId: paint.id },
-                      })
-                    }
-                    onClickRetweet={() => paintAction.onClickRetweet(paint.id)}
-                    onClickHeart={() => paintAction.onClickHeart(paint.id)}
-                    onClickViews={() => paintAction.onClickViews(paint.id)}
-                    onClickShare={() => paintAction.onClickShare(paint.id)}
-                    onClickMore={() => paintAction.onClickMore(paint.id)}
-                  />
-                ))}
+              <ContentLayout className="mt-0 pl-[12px] pr-[4px] pb-[50px] max-h-[calc(100%-94px)]">
+                <TimelineItemList list={paints} />
               </ContentLayout>
             ),
           },
           {
             label: '팔로우 중',
             content: (
-              <ContentLayout className="flex flex-col gap-[12px] pl-[12px] pr-[4px] mt-0 pb-[50px] max-h-[calc(100%-94px)] divide-y divide-blueGrey-400">
-                {[...paints].reverse().map((paint) => (
-                  <TimelineItemBox
-                    key={paint.id}
-                    item={paint}
-                    className="pt-[10px]"
-                    isShowMenu={
-                      paintAction.isShowMoreMenu.id === paint.id &&
-                      paintAction.isShowMoreMenu.show
-                    }
-                    onClickReply={() =>
-                      navigate({
-                        to: '/post/edit',
-                        search: { postId: paint.id },
-                      })
-                    }
-                    onClickRetweet={() => paintAction.onClickRetweet(paint.id)}
-                    onClickHeart={() => paintAction.onClickHeart(paint.id)}
-                    onClickViews={() => paintAction.onClickViews(paint.id)}
-                    onClickShare={() => paintAction.onClickShare(paint.id)}
-                    onClickMore={() => paintAction.onClickMore(paint.id)}
-                  />
-                ))}
+              <ContentLayout className="mt-0 pl-[12px] pr-[4px] pb-[50px] max-h-[calc(100%-94px)]">
+                <TimelineItemList list={[...paints].reverse()} />
               </ContentLayout>
             ),
           },
         ]}
         className="mt-[44px]"
-      />
-      <ReplyBottomSheet
-        id={paintAction.selectedPostId}
-        isOpen={paintAction.isBottomSheetOpen.reply}
-        onClose={() => paintAction.onCloseBottomSheet('reply')}
-      />
-      <ViewsBottomSheet
-        isOpen={paintAction.isBottomSheetOpen.views}
-        onClose={() => paintAction.onCloseBottomSheet('views')}
-      />
-      <ShareBottomSheet
-        id={paintAction.selectedPostId}
-        isOpen={paintAction.isBottomSheetOpen.share}
-        onClose={() => paintAction.onCloseBottomSheet('share')}
       />
     </>
   );
